利用 UDP 确定 IP 地址，与 Kotlin 保持一致

+++ MechDancer.Framework.Net/Remote/Modules/GroupMonitor.cs
@@ -18,7 +18,7 @@ namespace MechDancer.Framework.Net.Remote.Modules {
 			_broadcaster = Must<MulticastBroadcaster>(Host);
 		}
 
-		public void Yell() => _broadcaster.Value.Broadcast(UdpCmd.YellAsk);
+		public void Yell() => _broadcaster.Value.Broadcast((byte) UdpCmd.YellAsk);
 
 		public IReadOnlyCollection<byte> Interest => InterestSet;
 
@@ -30,7 +30,7 @@ namespace MechDancer.Framework.Net.Remote.Modules {
 					_detected?.Invoke(name);
 
 			if (cmd == (byte) UdpCmd.YellAsk) // 回应询问
-				_broadcaster.Value.Broadcast(UdpCmd.YellAck);
+				_broadcaster.Value.Broadcast((byte) UdpCmd.YellAck);
 		}
 
 		public override bool Equals(object obj) => obj is GroupMonitor;
